Enabled personalization and optimized route for premium users

+++ packages/db/src/graphql/schema/mutation/Trip/Trip.resolver.ts
@@ -32,6 +32,7 @@ const selectedFields = {
 };
 
 export const createTrip = async (
+  isPremium: boolean,
   userId: string,
   tripInput: CreateTripInput,
   tripPreferenceInput: CreateTripPreferenceInput,
@@ -49,6 +50,26 @@ export const createTrip = async (
   } = tripInput;
   try {
     const pois = await ctx.prisma.pointOfInterest.findMany({
+      where: {
+        OR: [
+          {
+            user: {
+              isNot: null,
+            },
+            businessPermit: {
+              isVerified: true,
+            },
+          }, // newly added place that is not yet verified must not be included
+          {
+            user: {
+              is: null,
+            },
+            businessPermit: {
+              is: null,
+            },
+          }, // include all scraped data
+        ],
+      },
       select: {
         ...selectedFields,
         accommodation: {
@@ -65,13 +86,16 @@ export const createTrip = async (
       (time: [number, number]) => getDesiredTravelHour(time),
     ) as number[];
 
-    const filteredPois = contentBasedFiltering(pois, tripPreferenceInput);
+    const pointOfInterests = isPremium
+      ? contentBasedFiltering(pois, tripPreferenceInput)
+      : pois;
 
-    console.log(filteredPois.length);
+    console.log(pointOfInterests.length);
 
     const suggestedItineraries = await generateItinerary(
+      isPremium,
       tripInput,
-      filteredPois,
+      pointOfInterests,
       duration,
       desiredTravelHours,
     );
@@ -80,10 +104,14 @@ export const createTrip = async (
       assignAccommodation(
         suggestedItineraries,
         tripInput,
-        filteredPois,
+        pointOfInterests,
         duration,
       ).map(async (itinerary) => {
-        const dailyItinerary = await createDailyItinerary(itinerary, tripInput);
+        const dailyItinerary = await createDailyItinerary(
+          isPremium,
+          itinerary,
+          tripInput,
+        );
         return dailyItinerary;
       }),
     );
@@ -102,15 +130,17 @@ export const createTrip = async (
         isTransportationIncluded: isTransportationIncluded,
         startingLocation: tripInput.startingLocation,
         timeSlots: tripInput.timeSlots as [number, number][],
-        tripPreference: {
-          create: {
-            accommodationType: tripPreferenceInput.accommodationType,
-            activities: tripPreferenceInput.activities,
-            amenities: tripPreferenceInput.amenities,
-            cuisines: tripPreferenceInput.cuisines,
-            diningStyles: tripPreferenceInput.diningStyles,
-          },
-        },
+        tripPreference: isPremium
+          ? {
+              create: {
+                accommodationType: tripPreferenceInput.accommodationType,
+                activities: tripPreferenceInput.activities,
+                amenities: tripPreferenceInput.amenities,
+                cuisines: tripPreferenceInput.cuisines,
+                diningStyles: tripPreferenceInput.diningStyles,
+              },
+            }
+          : undefined,
         dailyItineraries: {
           create: dailyItineraries.map((dailyItinerary, index) => ({
             dayIndex: index,
@@ -144,7 +174,7 @@ export const createTrip = async (
       },
     });
 
-    if (trip) {
+    if (trip && !isPremium) {
       await ctx.prisma.user.update({
         where: {
           id: userId,
@@ -165,40 +195,59 @@ export const createTrip = async (
 };
 
 export const deleteTrip = async (id: number, ctx: Context) => {
-  await ctx.prisma.tripPreference.delete({
-    where: {
-      tripId: id,
-    },
-  });
+  try {
+    const trip = await ctx.prisma.trip.findUnique({
+      where: {
+        id: id,
+      },
+      include: {
+        tripPreference: true,
+      },
+    });
 
-  await ctx.prisma.expense.deleteMany({
-    where: {
-      tripId: id,
-    },
-  });
+    if (trip?.tripPreference) {
+      await ctx.prisma.tripPreference.delete({
+        where: {
+          tripId: id,
+        },
+      });
+    }
 
-  await ctx.prisma.dailyItineraryPoi.deleteMany({
-    where: {
-      dailyItinerary: {
+    await ctx.prisma.expense.deleteMany({
+      where: {
         tripId: id,
       },
-    },
-  });
+    });
 
-  await ctx.prisma.dailyItinerary.deleteMany({
-    where: {
-      tripId: id,
-    },
-  });
+    await ctx.prisma.dailyItineraryPoi.deleteMany({
+      where: {
+        dailyItinerary: {
+          tripId: id,
+        },
+      },
+    });
 
-  return await ctx.prisma.trip.delete({
-    where: {
-      id: id,
-    },
-  });
+    await ctx.prisma.dailyItinerary.deleteMany({
+      where: {
+        tripId: id,
+      },
+    });
+
+    return await ctx.prisma.trip.delete({
+      where: {
+        id: id,
+      },
+    });
+  } catch (error) {
+    console.log(error);
+  }
 };
 
-export const regenerateTrip = async (id: number, ctx: Context) => {
+export const regenerateTrip = async (
+  isPremium: boolean,
+  id: number,
+  ctx: Context,
+) => {
   try {
     const pois = await ctx.prisma.pointOfInterest.findMany({
       select: {
@@ -240,6 +289,7 @@ export const regenerateTrip = async (id: number, ctx: Context) => {
     console.log(filteredPois.length);
 
     const suggestedItineraries = await generateItinerary(
+      isPremium,
       trip,
       filteredPois,
       duration,
@@ -253,7 +303,11 @@ export const regenerateTrip = async (id: number, ctx: Context) => {
         filteredPois,
         duration,
       ).map(async (itinerary) => {
-        const dailyItinerary = await createDailyItinerary(itinerary, trip);
+        const dailyItinerary = await createDailyItinerary(
+          isPremium,
+          itinerary,
+          trip,
+        );
         return dailyItinerary;
       }),
     );
